Various progress

+++ rpcs3/Loader/ELF32.h
@@ -15,7 +15,7 @@ struct Elf32_Ehdr
 	u16 e_type;
 	u16 e_machine;
 	u32 e_version;
-	u16 e_entry;
+	u32 e_entry;
 	u32 e_phoff;
 	u32 e_shoff;
 	u32 e_flags;
